[#6] refactor: 채팅방 관련(레포지토리) 추상화
